Simplify `View` (#10)
* View got way over-complicated; try a different approach

* RowKey

* Fix View<3> bug

* Opted against ComponentMapper

View Tables only need to be internally visible, so no wrapper is necessary

* Add readonly to containers

* Make Entity a readonly struct

* Readonly in builders

* Readony for View in Enumerator

* Add missing documentation comments.

* Remove RowKey

It didn't really simplify anything besides ViewRow, and then only slightly, and it complicated other things.

* Move Row creation to View

View is in a better place to optimize this, if/when we combine BindRow and Contains

* Simplify ViewEnumerator by moving row binding to View

ViewEnumerator no longer needs to care about the specific View.

Since View is a class type we can safely use interfaces for accessing it.

Further, this now allows View to have all the smarts caching indices while simultaneously
checking membership.

* View now calculates cached ViewRow indices

Minimal amount of reprocessing for View<2> and View<3>.

Slightly worse for View<1> since that's the one whose entities array we're iterating anyway. Could improve that by passing the entity index to the TryBindRow method but it'd only be really useful in View<1> due to the other optimization coming in.

* Iterate smallest number of entites in a View

Select the internal table with the fewest entities so that iteration has to walk fewer potential matches.